DT-6182 error handling, less unnecessary checks, renaming, formatting
