+++ app/component/itinerary/BicycleLeg.js
@@ -67,7 +67,6 @@ export default function BicycleLeg(
   const scooterSettingsOn = settings.scooterNetworks?.length > 0;
   const LOADSTATE = {
     UNSET: 'unset',
-    LOADING: 'loading',
     DONE: 'done',
   };
   const [nearestScooterState, setNearestScooterState] = useState({
@@ -177,7 +176,6 @@ export default function BicycleLeg(
     : leg.to.name;
 
   async function makeNearestScooterQuery(from) {
-    setNearestScooterState({ loading: LOADSTATE.LOADING });
     const planParams = {
       lat: from.lat,
       lon: from.lon,
@@ -188,39 +186,43 @@ export default function BicycleLeg(
       filterByPlaceTypes: ['VEHICLE_RENT'],
     };
 
-    const result = await fetchQuery(
-      relayEnvironment,
-      nearestQuery,
-      planParams,
-      {
-        force: true,
-      },
-    ).toPromise();
-    if (!result) {
-      setNearestScooterState({ loading: LOADSTATE.DONE });
-    } else {
-      const nearest = result?.viewer?.nearest?.edges;
-      // filter out the ones that are not scooters or from the same network as the original scooter
-      const filteredNearest = nearest
-        ?.filter(
-          n =>
-            n?.node?.place?.__typename === 'RentalVehicle' && // eslint-disable-line no-underscore-dangle
-            n?.node?.place?.rentalNetwork.networkId !== rentalVehicleNetwork,
-        )
-        // show only one scooter from each network
-        .filter(
-          (n, i, self) =>
-            self.findIndex(
-              t =>
-                t?.node?.place?.rentalNetwork.networkId ===
-                n?.node?.place?.rentalNetwork.networkId,
-            ) === i,
-        );
+    try {
+      const result = await fetchQuery(
+        relayEnvironment,
+        nearestQuery,
+        planParams,
+        {
+          force: true,
+        },
+      ).toPromise();
+      if (!result) {
+        setNearestScooterState({ loading: LOADSTATE.DONE });
+      } else {
+        const nearest = result.viewer.nearest.edges;
+        // filter out the ones that are not scooters or from the same network as the original scooter
+        const filteredNearest = nearest
+          ?.filter(
+            n =>
+              n.node.place.__typename === 'RentalVehicle' && // eslint-disable-line no-underscore-dangle
+              n.node.place.rentalNetwork.networkId !== rentalVehicleNetwork,
+          )
+          // show only one scooter from each network
+          .filter(
+            (n, i, self) =>
+              self.findIndex(
+                t =>
+                  t.node.place.rentalNetwork.networkId ===
+                  n.node.place.rentalNetwork.networkId,
+              ) === i,
+          );
 
-      setNearestScooterState({
-        nearest: filteredNearest,
-        loading: LOADSTATE.DONE,
-      });
+        setNearestScooterState({
+          nearest: filteredNearest,
+          loading: LOADSTATE.DONE,
+        });
+      }
+    } catch (error) {
+      setNearestScooterState({ nearest: [], loading: LOADSTATE.DONE });
     }
   }
